+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.3.0] - 2020-10-01
 ### Removed
 - Removed the `FormCreateMixin` (replaced with typed `createForm` function)
   - _The `FormCreaetMixin` (not truly a mixin...) did not work with TypeScript_